+++ R/rcax_superpops.R
@@ -8,6 +8,7 @@
 #'  `rcax_superpops()` will download 1000 records. Pass in `flist` to filter.
 #'  
 #' @export
+#' @param tablename The name of the table in the CAX API. See \code{vignette("CAX_Tables", package = "rCAX")}
 #' @template tableargs
 #' @template info
 #' @examples